Crate toreek_task_3

Modules

board

Contains board logic; accessing and moving pieces and initializing the board.

moves

Contains logic for how each chess piece should move.

piece

Contains definitions of a chess piece and related code.

position

Contains definition of a chess position and position logic.

tests 🔒

Structs

Game

Enums

GameState

Traits

GameTraits