toreek_task_3

Module board

Contains board logic; accessing and moving pieces and initializing the board.

Functions

check_castling 🔒

Checks if castling still is allowed. Checks if a Rook has moved, and from what Position, and if a king has moved and from what Position.

check_for_check 🔒

Checks if the opponent is in Check. See moves::king::king_illegal_moves.

check_for_promotion 🔒

Checks the board for a Pawn than can promote.

check_possible_moves 🔒

Checks if a Piece has any legal moves on a specific Position.

create_default_board 🔒

create_first_layer_piece_row 🔒

get_promotion_type 🔒

Gets the piece Type from a String. Returns None if the str does not contain a valid promotion type.

handle_castling 🔒

Checks if castling still is allowed. See check_castling. Also handles a castling move. Returns true if the function handles castling, as it overrides the default movement behavior.

handle_en_passant 🔒

Checks if a move is an en passant, or allows for an en passant next turn.

handle_in_progress 🔒

Checks if the move is legal, see Game::get_possible_moves. If true moves the piece, see handle_move and returns true, else returns false.

handle_move 🔒

Checks if an en passant was performed. See handle_en_passant. This move only removes an additional piece, so we can still move the piece normally. Checks if a castling was performed. See handle_castling. If true, we don’t want to move the piece normally, as castling provides a different movement pattern.

has_opponent_possible_moves 🔒

Recursively goes through every possible move for every Piece on the board for the opposite player. Checks if the opponent can do any legal moves. Used to determinate Checkmate or Stalemate.

post_turn_check 🔒

Checks various thing right after a turn.