toreek_task_3::board

Function check_possible_moves

pub(crate) fn check_possible_moves(
    position: &Position,
    game: &Game,
    piece: &Piece,
) -> Vec<String>

Checks if a Piece has any legal moves on a specific Position.

Gets all moves a Piece can do, see Piece::get_possible_moves for per Type logic.

Checks if any moves puts the own King into check. See handle_move and check_for_check. Moves that puts the own King into check are considered illegal and are not added to the final Vec<String> of allowed_moves. Returns a Vec<String> of all legal moves a Piece can do.