toreek_task_3::moves

Module king

Contains the logic behind the move pattern for type King. See [king_possible_moves] for entry point.

Functions

king_castling 🔒

Handles castling when a King is selected. Just an if soup, nothing of interest to see here.

king_illegal_moves 🔒

Returns a Vec<String> of every possible move the opposite player can do, that can threaten another piece. Observe that a Pawn, unlike other [Piece]s can move forwards, but can only threaten pieces diagonally.

king_moves 🔒

Returns a 3x3 grid around the Position of the King, excluding the position in the middle. Excludes invalid positions, see add_position_if_valid for details.

king_possible_moves 🔒

Entry point for King move pattern logic.