toreek_task_3::moves::king

Function king_illegal_moves

pub(crate) fn king_illegal_moves(
    game: &Game,
    checked_color: &Colour,
) -> Vec<String>

Returns a Vec<String> of every possible move the opposite player can do, that can threaten another piece. Observe that a Pawn, unlike other [Piece]s can move forwards, but can only threaten pieces diagonally.

Loops every Position of crate::Game::board. Checks for [Piece] of opposite Colour.

If the [Piece] is a Pawn check the positions diagonally in front of the Pawn. We can’t reuse [Piece::get_possible_moves] as the default move pattern doesn’t necessarily threaten other pieces. If the [Piece] is any other piece, get their possible moves from [Piece::get_possible_moves].