toreek_task_3

Module piece

Contains definitions of a chess piece and related code.

Structs

Piece

A struct containing a representation of a chess piece. Holds a Type e.g. Knight. Holds a Colour e.g. White.

Enums

Colour

Type